rename package name in java files
