+++ kernel-spark/src/main/java/io/delta/kernel/spark/read/SparkPartitionReader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.dsv2.read;
+package io.delta.kernel.spark.read;
 
 import java.io.IOException;
 import org.apache.spark.sql.catalyst.InternalRow;
