+++ kernel-spark/src/main/java/io/delta/kernel/spark/read/SparkScan.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.dsv2.read;
+package io.delta.kernel.spark.read;
 
 import io.delta.kernel.data.MapValue;
 import io.delta.kernel.data.Row;
