+++ kernel-spark/src/main/java/io/delta/kernel/spark/utils/ScalaUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.dsv2.utils;
+package io.delta.kernel.spark.utils;
 
 import java.util.Map;
 import scala.Tuple2;
