+++ kernel-spark/src/main/java/io/delta/kernel/spark/utils/SerializableKernelRowWrapper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.dsv2.utils;
+package io.delta.kernel.spark.utils;
 
 import io.delta.kernel.data.Row;
 import io.delta.kernel.defaults.internal.json.JsonUtils;
