+++ kernel-spark/src/test/java/io/delta/kernel/spark/Dsv2BasicTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.dsv2;
+package io.delta.kernel.spark;
 
 import static org.junit.jupiter.api.Assertions.*;
