+++ kernel-spark/src/test/java/io/delta/kernel/spark/SparkDsv2TestBase.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.dsv2;
+package io.delta.kernel.spark;
 
 import io.delta.kernel.defaults.engine.DefaultEngine;
 import io.delta.kernel.engine.Engine;
