+++ kernel-spark/src/test/java/io/delta/kernel/spark/catalog/TestCatalog.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.dsv2.catalog;
+package io.delta.kernel.spark.catalog;
 
 import io.delta.kernel.Operation;
 import io.delta.kernel.defaults.engine.DefaultEngine;
