+++ kernel-spark/src/test/java/io/delta/kernel/spark/read/SparkScanBuilderTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.delta.spark.dsv2.read;
+package io.delta.kernel.spark.read;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
